Update script.js
Added math for generator that establishes a for loop with variable i that equals 0. It will continue running the loop with each iteration incrementing by 1 while i is less than the passwordLength.

Established variable inside loop randomPassword to run a math.random times the available character length specified. That number is then rounded down to the nearest floor number.

Then established a password with an addition assignment to availableChars appended to password so every iteration of the for loop a random character is selected and added to the availableChars string.

The loop continues until i is the same value as passwordLength.

Then it returns the password generated.